Removed metro timer for the moment

+++ Universal_Serial_Adapter/Config.cpp
@@ -46,7 +46,7 @@ timeout Config::getTimeout() {
 	return currentTimeout;
 }
 
-long Config::getTimeoutMilis() {
+int Config::getTimeoutMilis() {
 	switch (currentTimeout) {
 		case 0: // tenseconds
 			return 10000;
